requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
openai>=1.0.0,<2.0.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiosqlite>=0.17.0
greenlet>=2.0.0
click>=8.0.0
rich>=10.0.0
dateparser>=1.1.0
